GUI Start Pause Stop Button Fixed
Updated that GameController.py so the actual buttons: start, pause, and stop all work as intended. Note: after pausing and resuming it will start the dungeon floor over, it is not paused middle of the dungeon routine. Also added some debugging options, currently on, set to false if you don't want to see all the stuff in the console.