Enable dungeons 1-15 Tested Working
Added selections for the first 15 dungeons, but honestly any dungeon is playable. Dungeons 3, 6, and 10 have very specific custom paths, but anything else all runs the 'powerplay' path that should get through 99% of maps. There is a problem watch out for dungeons with spikes and saw blades on the floors as they will not be avoided. I tested dungeons 8 and 11 myself and worked very well. 

Note that when playing any dungeon besides 3, 6, and 10 the levels in the GUI will not match the level of the room you are in. It will play continuously until the endgame screen is reached either you die or win! You can farm any dungeon, give it a try.