Fixed Very Rare Loop Bug
The game would get stuck in the energy loop IF the game crashed while on level 0.  Fixed the problem!  The game will now continue if an in_game state is detected while on level zero. All other levels 1+ already have this protection included; if the game crashes on any level now it should not get stuck on the detect energy phase.